backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
langchain==0.0.350
langchain-community==0.0.1
chromadb==0.4.17
openai==1.3.7
pypdf2==3.0.1
python-docx==1.1.0
pandas==2.1.4
numpy==1.25.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
sentence-transformers==2.2.2
torch==2.1.1
transformers==4.36.2
httpx==0.25.2
aiofiles==23.2.0
